+++ backend/src/main/java/dev/tripdraw/file/application/FileUrlMaker.java
@@ -1,4 +1,4 @@
-package dev.tripdraw.application.file;
+package dev.tripdraw.file.application;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
