+++ backend/src/main/java/dev/tripdraw/file/exception/FileIOExceptionType.java
@@ -1,4 +1,4 @@
-package dev.tripdraw.exception.file;
+package dev.tripdraw.file.exception;
 
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
