+++ backend/src/test/java/dev/tripdraw/file/application/FileUrlMakerTest.java
@@ -1,4 +1,4 @@
-package dev.tripdraw.application.file;
+package dev.tripdraw.file.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
